Pass individual informers, move DRA controllers to resource.go, simplify retry logic and metric tests
